+++ src/ResourceManager/Profile/Commands.Profile/Models/RMProfileClient.cs
@@ -68,7 +68,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Common
                 var token = AcquireAccessToken(account, environment, tenantId, password, promptBehavior);
                 if(TryGetTenantSubscription(token, account, environment, tenantId, subscriptionId, subscriptionName, out newSubscription, out newTenant))
                 {
-                    account.SetProperty(AzureAccount.Property.Tenants, new[] { newTenant.Id.ToString() });
+                    account.SetOrAppendProperty(AzureAccount.Property.Tenants, new[] { newTenant.Id.ToString() });
                 }
             }
             // (tenant is not provided and subscription is present) OR
@@ -240,7 +240,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Common
 
         public List<AzureTenant> ListTenants(string tenant)
         {
-            return ListAccountTenants(_profile.Context.Account, _profile.Context.Environment, null, ShowDialog.Auto)
+            return ListAccountTenants(_profile.Context.Account, _profile.Context.Environment, null, ShowDialog.Never)
                 .Where(t => tenant == null ||
                             tenant.Equals(t.Id.ToString(), StringComparison.OrdinalIgnoreCase) ||
                             tenant.Equals(t.Domain, StringComparison.OrdinalIgnoreCase))
@@ -380,13 +380,16 @@ namespace Microsoft.Azure.Commands.ResourceManager.Common
             {
                 try
                 {
-                    subscriptions.AddRange(ListSubscriptions(tenant.Id.ToString()));
+                    subscriptions.AddRange(
+                        ListSubscriptions(
+                            (tenant.Id == Guid.Empty) ? tenant.Domain:tenant.Id.ToString()));
                 }
                 catch (AadAuthenticationException)
                 {
-                    WriteWarningMessage(string.Format("Could not authenticate user account {0} with tenant {1}.  " +
-                       "Subscriptions in this tenant will not be listed. Please login again using Login-AzureRmAccount " +
-                       "to view the subscriptions in this tenant.", _profile.Context.Account, tenant));
+                    WriteWarningMessage(string.Format(
+                        Microsoft.Azure.Commands.Profile.Properties.Resources.UnableToLogin, 
+                        _profile.Context.Account, 
+                        tenant));
                 }
 
             }
@@ -481,19 +484,24 @@ namespace Microsoft.Azure.Commands.ResourceManager.Common
                     }
                     else
                     {
-                        var subscriptions = subscriptionClient.Subscriptions.List().Subscriptions;
-                        if (subscriptions != null && subscriptions.Any())
+                        var subscriptions = (subscriptionClient.Subscriptions.List().Subscriptions ?? 
+                                                new List<Microsoft.Azure.Subscriptions.Models.Subscription>())
+                                            .Where(s => "enabled".Equals(s.State, StringComparison.OrdinalIgnoreCase) ||
+                                                        "warned".Equals(s.State, StringComparison.OrdinalIgnoreCase));
+
+                        if (subscriptions.Any())
                         {
                             if (subscriptionName != null)
                             {
-                                subscriptionFromServer = subscriptions.FirstOrDefault(s => s.DisplayName.Equals(subscriptionName, StringComparison.OrdinalIgnoreCase));
+                                subscriptionFromServer = subscriptions.FirstOrDefault(
+                                    s => s.DisplayName.Equals(subscriptionName, StringComparison.OrdinalIgnoreCase));
                             }
                             else
                             {
-                                if (subscriptions.Count > 1)
+                                if (subscriptions.Count() > 1)
                                 {
                                     WriteWarningMessage(string.Format(
-                                        "TenantId '{0}' contains more than one subscription. First one will be selected for further use. " +
+                                        "TenantId '{0}' contains more than one active subscription. First one will be selected for further use. " +
                                         "To select another subscription, use Set-AzureRmContext.",
                                         tenantId));
                                 }
@@ -569,14 +577,30 @@ namespace Microsoft.Azure.Commands.ResourceManager.Common
                 {
                     result =
                         account.GetPropertyAsArray(AzureAccount.Property.Tenants)
-                            .Select( ti => new AzureTenant() 
+                            .Select(ti =>
                             {
-                                Id = new Guid(ti),
-                                Domain = AccessTokenExtensions.GetDomain(account.Id)
+                                var tenant = new AzureTenant();
+
+                                Guid guid;
+                                if (Guid.TryParse(ti, out guid))
+                                {
+                                    tenant.Id = guid;
+                                    tenant.Domain = AccessTokenExtensions.GetDomain(account.Id);
+                                }
+                                else
+                                {
+                                    tenant.Domain = ti;
+                                }
+
+                                return tenant;
                             }).ToList();
                 }
-                
-             }
+                if(!result.Any())
+                {
+                    throw;
+                }
+
+            }
 
             return result;
         }
@@ -608,7 +632,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Common
                         subscriptions.Subscriptions.Select(
                             (s) =>
                                 s.ToAzureSubscription(new AzureContext(_profile.Context.Subscription, account,
-                                    environment, CreateTenantFromString(tenantId))));
+                                    environment, CreateTenantFromString(tenantId, accessToken.TenantId))));
                 }
 
                 return new List<AzureSubscription>();
@@ -622,8 +646,8 @@ namespace Microsoft.Azure.Commands.ResourceManager.Common
                 WarningLog(message);
             }
         }
-
-        private static AzureTenant CreateTenantFromString(string tenantOrDomain)
+        
+        private static AzureTenant CreateTenantFromString(string tenantOrDomain, string accessTokenTenantId)
         {
             AzureTenant result = new AzureTenant();
             Guid id;
@@ -633,6 +657,7 @@ namespace Microsoft.Azure.Commands.ResourceManager.Common
             }
             else
             {
+                result.Id = Guid.Parse(accessTokenTenantId);
                 result.Domain = tenantOrDomain;
             }