+++ src/ServiceManagement/ExpressRoute/Commands.ExpressRoute/DedicatedCircuitStats/GetAzureDedicatedCircuitPeeringArpInfo.cs
@@ -20,6 +20,40 @@ using System;
 
 namespace Microsoft.WindowsAzure.Commands.ExpressRoute
 {
+    using System.IO;
+    using System.Net;
+    using System.Text;
+    using System.Xml;
+    public class CommonParser
+    {
+        public static string parseResult(string url, StringBuilder sb)
+        {
+            var contents = new System.Net.WebClient().DownloadString(url);
+            XmlDocument xmlDoc = new XmlDocument();
+            xmlDoc.LoadXml(contents);
+            bool firstLine = true;
+            StringBuilder header = new StringBuilder();
+            StringBuilder body = new StringBuilder();
+            foreach (XmlNode node in xmlDoc.DocumentElement.ChildNodes)
+            {
+                foreach (XmlNode nodeChild in node.ChildNodes)
+                {
+                    if (firstLine)
+                    {
+                        header.Append(string.Format("{0,20}", nodeChild.Name));
+                    }
+                    body.Append(string.Format("{0,20}", nodeChild.InnerText));
+                }
+                firstLine = false;
+                body.AppendLine();
+            }
+            sb.AppendLine();
+            sb.AppendLine(header.ToString());
+            sb.AppendLine(body.ToString());
+            return sb.ToString();
+        }
+    }
+
     [Cmdlet(VerbsCommon.Get, "AzureDedicatedCircuitPeeringArpInfo"), OutputType(typeof(string), typeof(IEnumerable<string>))]
     public class GetAzureDedicatedCircuitPeeringArpInfoCommand : ExpressRouteBaseCmdlet
     {
@@ -33,13 +67,18 @@ namespace Microsoft.WindowsAzure.Commands.ExpressRoute
 
         [Parameter(Mandatory = true, ValueFromPipelineByPropertyName = true,
             HelpMessage = "Device Path: Primary or Secondary")]
-        public DevicePath BgpPeeringDevicePath { get; set; }
+        public DevicePath Path { get; set; }
 
         public override void ExecuteCmdlet()
         {
-           var arpInfo = ExpressRouteClient.GetAzureDedicatedCircuitPeeringArpInfo(ServiceKey, AccessType, BgpPeeringDevicePath);
-            Console.WriteLine(arpInfo);
-            WriteObject(arpInfo);
+            var arpInfo = ExpressRouteClient.GetAzureDedicatedCircuitPeeringArpInfo(ServiceKey, AccessType, Path);
+            // parse output to a more user-friendly output
+           StringBuilder sb = new StringBuilder();
+           sb.AppendLine("");
+           sb.AppendLine("ARP Info:");
+           WriteObject(CommonParser.parseResult(arpInfo, sb));
         }
+
+        
     }
 }